fix: CI errors — unused imports, btts_pct float type, dead JoueursTab
- Remove unused imports (MODEL_FEATURES, time, boxscoresummaryv2)
- btts_pct/over25_pct: int -> float (fixes pydantic validation skip)
- Remove unused JoueursTab component (frontend build error)

+++ src/api/schemas.py
@@ -554,10 +554,10 @@ class AIScanMatch(BaseModel):
     away_xg_avg: float | None = None
     home_clean_sheets: int | None = None
     away_clean_sheets: int | None = None
-    home_btts_pct: int | None = None
-    away_btts_pct: int | None = None
-    home_over25_pct: int | None = None
-    away_over25_pct: int | None = None
+    home_btts_pct: float | None = None
+    away_btts_pct: float | None = None
+    home_over25_pct: float | None = None
+    away_over25_pct: float | None = None
     home_corners_avg: float | None = None
     away_corners_avg: float | None = None
     home_cards_avg: float | None = None
